+++ InventoryManagement/SmartPickup/SmartPickup.cs
@@ -33,7 +33,7 @@ public sealed class SmartPickup : ModSystem {
         EmitSmartPickup(cursor, newItem, (Player self, Item item, GetItemSettings settings) => {
             if (vanillaGetItem) return item;
             if (Configs.SmartPickup.RefillMouse) item = SmartEquip.RefillMouse(self, item, settings);
-            if (Configs.PreviousSlot.Enabled) item = PreviousSlot.PickupItemToPreviousSlot(self, item, settings);
+            if (Configs.SmartPickup.PreviousSlot) item = PreviousSlot.PickupItemToPreviousSlot(self, item, settings);
             return item;
         });
     }
@@ -53,9 +53,10 @@ public sealed class SmartPickup : ModSystem {
         // ++<upgradeItems>
         EmitSmartPickup(cursor, newItem, (Player self, Item item, GetItemSettings settings) => {
             if (vanillaGetItem || settings.NoText) return item;
-            if (!item.IsAir && Configs.UpgradeItems.Enabled) item = SmartEquip.UpgradeItems(self, item, settings);
+            if (!item.IsAir && Configs.SmartPickup.QuickStack) item = SmartEquip.QuickStack(self, item, settings);
+            if (!item.IsAir && Configs.SmartPickup.UpgradeItems) item = SmartEquip.UpgradeItems(self, item, settings);
             if (!item.IsAir && Configs.SmartPickup.AutoEquip) item = SmartEquip.AutoEquip(self, item, settings);
-            if (!item.IsAir && Configs.SmartPickup.VoidBag) item = SmartEquip.VoidBag(self, item, settings);
+            if (!item.IsAir && Configs.SmartPickup.VoidBagFirst) item = SmartEquip.VoidBagFirst(self, item, settings);
             return item;
         });
     }